buffer: add advance(unsigned) back
advance(unsigned) and seek(int) were replaced with advance(size_t> and
seek(ssize_t) in af631ef. this change broke the backward compatibility
of librados/librbd's ABI.

- deprecates advance(unsigned) and seek(int)
- update the callers to use the prefered versions.

Fixes: http://tracker.ceph.com/issues/17809

+++ src/common/buffer.cc
@@ -1082,7 +1082,7 @@ static simple_spinlock_t buffer_debug_lock = SIMPLE_SPINLOCK_INITIALIZER;
     : iterator_impl<is_const>(i.bl, i.off, i.p, i.p_off) {}
 
   template<bool is_const>
-  void buffer::list::iterator_impl<is_const>::advance(ssize_t o)
+  void buffer::list::iterator_impl<is_const>::advance(int o)
   {
     //cout << this << " advance " << o << " from " << off << " (p_off " << p_off << " in " << p->length() << ")" << std::endl;
     if (o > 0) {
@@ -1121,7 +1121,7 @@ static simple_spinlock_t buffer_debug_lock = SIMPLE_SPINLOCK_INITIALIZER;
   }
 
   template<bool is_const>
-  void buffer::list::iterator_impl<is_const>::seek(size_t o)
+  void buffer::list::iterator_impl<is_const>::seek(unsigned o)
   {
     p = ls->begin();
     off = p_off = 0;
@@ -1313,12 +1313,12 @@ static simple_spinlock_t buffer_debug_lock = SIMPLE_SPINLOCK_INITIALIZER;
     : iterator_impl(l, o, ip, po)
   {}
 
-  void buffer::list::iterator::advance(ssize_t o)
+  void buffer::list::iterator::advance(int o)
   {
     buffer::list::iterator_impl<false>::advance(o);
   }
 
-  void buffer::list::iterator::seek(size_t o)
+  void buffer::list::iterator::seek(unsigned o)
   {
     buffer::list::iterator_impl<false>::seek(o);
   }
